+++ abilian/services/activity/__init__.py
@@ -1,2 +1,4 @@
 from .service import ActivityService
 from .models import ActivityEntry
+
+__all__ = ['ActivityEntry', 'ActivityService']