+++ abilian/web/csrf.py
@@ -1,8 +1,8 @@
 # coding=utf-8
-from functools import wraps
-from datetime import timedelta
-from werkzeug.exceptions import Forbidden
 
+from functools import wraps
+
+from werkzeug.exceptions import Forbidden
 from flask import Blueprint, jsonify, current_app
 from flask.ext.wtf import Form